+++ crates/rangelist/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+- Add `Rangelist::set_lower_bound` and `Rangelist::set_upper_bound` methods to
+  tighten the lower and upper bounds of a `Rangelist`.
+
 ## [0.3.1] - 2025-08-13
 
 ### Fixed